Moving size to a mixin for code reuse
